requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.9
celery>=5.3.6
redis>=5.0.1
python-dotenv>=1.0.0
google-api-python-client>=2.0.0
google-auth-oauthlib>=1.0.0
youtube-transcript-api>=0.6.1
aiohttp>=3.8.1